drm/msm/mdp5: Prevent calling vblank calculation funcs when crtc is disabled
The drm_driver ops get_vblank_timestamp and get_vblank_counter read MDP5
registers to assist in vblank calculations. When a CRTC is disabled, it's
possible that the MDP5 clocks are disabled, and accessing these registers
should be avoided. Calculating vblank count or scanout position anyway
don't make much sense either when a CRTC is disabled.

Change-Id: If15c9a57b5b08804c9fc19ae14e514588ddb7b02
Signed-off-by: Archit Taneja <architt@codeaurora.org>